Corrected links to original def of describedby
Changed link to POWDER-DR document to link to the appendix where the original (correct) definiton of the describedby link relationship. The text in the main part of the doc is recorded as being in error at http://www.w3.org/2007/powder/powder-errata#describedby. Also, the title of the POWDER doc in the right hand column was linked to the CORS spec by mistake.