One header from RFC 4229
The Title header field is sometimes useful to carry information.  The RFC doesn't explicitly define what the title field does, it only references a document by @timbl back in 1994, and I grandfathered in the text from there. Is this right @dret?

Probably there are more headers in RFC 4229 that should be added, but one would have to subtract the things that are defined elsewhere.  What happens if there are two claims to define a single web concept?