adding link tool tips for #46 (maybe that's all it takes)
this took much longer than it should have. MD is horrible. in this
case, escaping '<' in link tool tips seems to be necessary in jekyll,
even though it is not necessary in atom. it would be good to have a
proper function for MD output that handles all necessary escaping.